Update email validation I changed the validation code in signup.js
Originally it was setting the component 'email' state based on the string input by the user, THEN it would validate that input string itself. This would work in the majority of cases, but due to the async nature of setState in React it could lead to a discrepancy wherein bad data is allowed to pass through.

The way it works now is as follows:
1. The user inputs a string into the email field
2. setEmail() is called, and the component state updates
3. After the component state updates, the useLayoutEffect() hook is called and validates the 'email' state value itself, rather than the string input by the user.

I used 'useRef' and 'useLayoutEffect' instead of a more simple 'useEffect' hook because I didn't want the form to immediately throw an error on initial render. So, the useRef/useLayoutEffect assure that the error will not be thrown when the component loads for the first time on the page.